Revert "Move async logic into worker"
This reverts commit 6af1281d5707a9f37bd9ab40ef89498ad40ab18f.
